backend/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
google-generativeai==0.8.3
python-dotenv==1.0.0
requests==2.32.4
gunicorn==23.0.0
supabase==2.11.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1